pyproject.toml
[project]
name = "udesc-testes-de-software-trabalho-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.3",
    "httpx>=0.27.2",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "python-jose>=3.3.0",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]